+++ lib/trailblazer/activity/dsl/linear/normalizer/inherit.rb
@@ -13,17 +13,17 @@ module Trailblazer
 
             # Options you want to have stored and inherited can be
             # declared using Record.
-            Record = Struct.new(:options, :type, :non_symbol_options?) # FIXME: i hate symbol vs. non-symbol.
+            Record = Struct.new(:options, :type)
 
-            def Record(options, type:, non_symbol_options: true)
-              {Record.new(options, type, non_symbol_options) => nil}
+            def Record(options, type:)
+              {Record.new(options, type) => nil}
             end
 
             # Currently, the {:inherit} option copies over {:extensions} from the original step and merges them with new :extensions.
             #
             ### Recall
             # Fetch remembered options and add them to the processed options.
-            def recall_recorded_options(ctx, non_symbol_options:, sequence:, id:, inherit: nil, extensions: [], **)
+            def recall_recorded_options(ctx, sequence:, id:, inherit: nil, **)
               return unless inherit === true || inherit.is_a?(Array)
 
               # E.g. {variable_mapping: true, wiring_api: true}
@@ -37,24 +37,15 @@ module Trailblazer
 
               row = find_row(sequence, id) # from this row we're inheriting options.
 
-              # DISCUSS: should we maybe revert the idea of separating options by type?
-              # Anyway, key idea here is that Record() users don't have to know these details
-              # about symbol vs. non-symbol.
-              symbol_options_to_merge     = {}
-              non_symbol_options_to_merge = {}
-
+              # recorded_options: for example [:fast_track, :custom_output_tuples, :extensions]
               row.data[:recorded_options].each do |type, record|
                 next unless types_to_recall[type]
 
-                target = record.non_symbol_options? ? non_symbol_options_to_merge : symbol_options_to_merge
-                target.merge!(record.options)
+                ctx = record.options.merge(ctx)
               end
 
-              ctx[:non_symbol_options] = non_symbol_options_to_merge.merge(non_symbol_options)
 
-              ctx.merge!(symbol_options_to_merge)
-
-              ctx.merge!(
+              ctx.merge(
                 inherited_recorded_options: row.data[:recorded_options]
               )
             end
@@ -67,19 +58,19 @@ module Trailblazer
             ### Record
             # Figure out what to remember from the options and store it in {row.data[:recorded_options]}.
             # Note that this is generic logic not tied to variable_mapping, OutputTuples or anything.
-            def compile_recorded_options(ctx, non_symbol_options:, **)
+            def compile_recorded_options(ctx, **)
               recorded_options = {}
 
-              non_symbol_options
+              ctx
                 .find_all { |k, v| k.instance_of?(Record) }
                 .collect  do |k, v|
                   recorded_options[k.type] = k   # DISCUSS: we overwrite potential data with same type.
                 end
 
-              ctx.merge!(
+              ctx.merge(
                 recorded_options:   recorded_options,
                 # add {row.data[:recorded_options]} in Sequence:
-                non_symbol_options: non_symbol_options.merge(Strategy.DataVariable() => :recorded_options)
+                Strategy.DataVariable() => :recorded_options
               )
             end
           end # Inherit